refactor(infra): rename app_container module to containers
